Add a dialog that allows the user to choose styles
This may be easier for some users than having to look up and type in
the style ids.

The default styles are downloaded from the Building Styles Master IID
public Google Spreadsheet.
These styles are caches locally in a BuildingStyleCache.json file in
the same folder as the application.
The user can force an update from the Google Spreadsheet by deleting
that cache file, but it isn't available as a button in the UI due to
such a feature removing any changes the user made to the style list.